API Move fieldset-related API into new trait
CompositeField has API that lets it change between a `div` and a HTML
`fieldset`. Moving this into a trait lets other fields behave the same
way without having to inherit everything about CompositeField.
