+++ src/Forms/CompositeField.php
@@ -15,6 +15,8 @@ use SilverStripe\Dev\Debug;
  */
 class CompositeField extends FormField
 {
+    use FieldSetTrait;
+
     private static array $field_validators = [
         CompositeFieldValidator::class,
     ];
@@ -40,18 +42,6 @@ class CompositeField extends FormField
      * count of your $children.
      */
     protected $columnCount = null;
-
-    /**
-     * @var string custom HTML tag to render with, e.g. to produce a <fieldset>.
-     */
-    protected $tag = 'div';
-
-    /**
-     * @var string Optional description for this set of fields.
-     * If the {@link $tag} property is set to use a 'fieldset', this will be
-     * rendered as a <legend> tag, otherwise its a 'title' attribute.
-     */
-    protected $legend;
 
     protected $schemaDataType = FormField::SCHEMA_DATA_TYPE_STRUCTURAL;
 
@@ -162,43 +152,6 @@ class CompositeField extends FormField
         $this->children = $children;
         $children->setContainerField($this);
         return $this;
-    }
-
-    /**
-     * @param string $tag
-     * @return $this
-     */
-    public function setTag($tag)
-    {
-        $this->tag = $tag;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTag()
-    {
-        return $this->tag;
-    }
-
-    /**
-     * @param string $legend
-     * @return $this
-     */
-    public function setLegend($legend)
-    {
-        $this->legend = $legend;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getLegend()
-    {
-        return $this->legend;
     }
 
     public function extraClass()
